calendar-to-asana/requirements.txt
python-dotenv>=1.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0
asana>=0.21.0
